+++ XADMaster/XADLZXParser.m
@@ -12,7 +12,7 @@
 +(BOOL)recognizeFileWithHandle:(CSHandle *)handle firstBytes:(NSData *)data name:(NSString *)name
 {
 	const uint8_t *bytes=[data bytes];
-	int length=[data length];
+	NSInteger length=[data length];
 
 	return length>=10&&bytes[0]=='L'&&bytes[1]=='Z'&&bytes[2]=='X';
 }
